+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.5.1] - 2026-07-15
+- Clean git history
+
 ## [0.5.0] - 2026-07-15
 
 ### Added